Add ability to consume data from Monobank API by time ranges
Following was done:
* time_helper was added
* request validation was added
